Adds battery charging circuit with an MCP73831
The battery will be 400mA, which means we should use 200mA to charge the
battery. This would mean a 5kΩ prog resistor. A 5.1kΩ resistor was
chosen instead to simplify the BOM since we'll already use some of those
in the USB configuration. This means a charging current of approximately
196mA.