Adds voltage divider to sense battery level
I decided to go with a simple voltage divider instead of a more complex
circuit that could be turned off because with two 100k resistors this
should mean a 21µA current when the battery is full (at 4.2V). This
seems like a good tradeoff to have a simpler circuit/BOM and layout
(since space is important here). Using two 100k might make the ADC a bit
noisy here, but this should be enough to have a good idea of the battery
level.

I'm using IO35 as that is in the first ADC block and block 2 is used by
wifi[1].

[1]: https://docs.micropython.org/en/latest/esp32/quickref.html#adc-analog-to-digital-conversion