Removes IMU
This PCB will be the base, on which we will connect other PCBs that add
functionality to it. The IMU will be on it's own PCB, with it's own LED.
The keypad will be a third PCB with the buttons and LEDs. The base and
the add on PCB will be connected by a FFC connector that breaks out the
the GPIO pins we're not using on the base.