Moves ID latch/clock/data pins
This means some of the IO that are broken out to the add-on changed, but
not only are new new pins (GPIO/14,12,13) better placed on the module,
but they also are block2 of the ADC, which means they can't be used for
analog readings while wifi is used. If we're going to use a block of
pins for digital IO, might as well use these.