Uses the MCU to check the battery charging status
The STAT pin on the MCP73831 is low when the battery is charging,
high when it's done and hiigh Z when shutdown or the battery is not
present. This means we can read this pin from the microcontroller to
know if the battery is being charged and indicate that using the RGB Led
already present for battery level indication. The pull-up is necessary
since when the charger or battery are not present the pin would be
floating.

We used GPIO39 on the esp32 for this reading, which means this pin is no
longer broken out to the add-on boards. The connector changed to
accomodate for that.