Adds BNO055 sensor to schematics
The unused pins are disconnected for now. If we have pins left after
adding other components, we can connect them.

3v3 is connected to the board through a jumper that's normally closed to
Vin since that should be the pin used. In the prototype, 3vo was used
and worked. After we confirm the PCB works with 3v3 connected to Vin as
expected, we can remove this jumper.