Adds unused pins to the add-on connector
Those pins were used to get the ID from the DIP switches, but since we
don't have them anymore, we can use them as GPIOs.

This commit changes the connector on the baseboard (including rerouting
the traces) and updates the add-on connector schematic symbol in the
symbol library, but does not update yet the boards that use the add-on
connector (sensor board and breakout board).