Fixes orientation for sensor board
The idea is to use the base board with the ESP32 facing down, and the
sensorboard with the sensor facing up. That way the connectors will face
each other correctly and the sensor will have space above (it wouldn't
fit between the boards because the battery will already be there).