Fixes issue with battery status input
The battery status will be pulled high by the MCP73831 to it's VDD (5V)
when charging is done. This is higher than the 3.3V (+0.3V) allowed as
input to the ESP32. By adding a voltage divider we bring this to a lower
level with which we can interface to the ESP32.

When no 5V or no battery is present, the pin will be floating, so this
pin should have a internal pull-up to make sure we read the value
correctly. An internal pull-up will mean we'll see a high value with any
of charge complete, no 5V input, or no battery, which means a high value
whenever not charging.

The 150k resistor adds to the BOM, but in conjunction with the 100k it
will bring the 5V down to 3V, which is a safe value to input to the
ESP32.