Adds box to hold the hardware during performances
This was designed out of this repo, so there are no step by step
commits, sorry.

The main idea is that the base board sits on the stands (the ones with
holes should be fitted with heat-set inserts to accept M2 spacers) with
the ESP32 facing down, the add-on sits on top of stands that separate
the two PCBs, and the battery in between (ideally with some foam to hold
it in place and avoid damage). In my preliminary tests, the battery
between the boards did not significantly reduce the antenna's
performance. The big gap on the right is to accommodate the FFC cables,
which we couldn't find in smaller size.

Different lids should be designed for each add-on type.