Generic lid prototype only sized to the box and with press fit walls
It's already clear that a lot of the variables in common.scad will need
to go in the respective files, but that's for later.

The models allows for a screw hole to be placed, so one of the screws
can hold the lid. The reason we only set one screw is that the other one
is needed to hold the PCB close to the spacer, which makes it stay in
place much better (which is needed for the keypad, for example). The
default is to not have the hole, as the press fit already works quite
well.

There's also a flag to include the logo on the lid or not. It's off by
default because the only sensible way to print this model is with the
top of the lid facing down, which means there's support needed there
which makes adhesion of the first layer much harder and also is not very
easy to clean up.